# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DAM Platform API
  version: 1.0.0
  description: |
    REST API of the DAM platform: accounts, asset browsing, direct-to-storage
    multipart uploads, and admin reporting.

    **Base URL.** Through the web container's nginx the API lives under `/api`
    (`http://localhost:8080/api`), and nginx strips that prefix. The API
    container also publishes port 3000 directly. Resource paths are versioned
    under `/v1`. `/health` and `/docs` are not versioned.

    **Authentication.** Register and login set an HttpOnly cookie named
    `dam_token`. Page scripts cannot read it. The JSON body contains the user
    and not the token. The browser sends the cookie on later requests.
    `SameSite=Strict` keeps other sites from attaching it. A non-browser client
    may instead send `Authorization: Bearer <token>`. Everything except
    `/health/*`, `/docs`, `/v1/auth/register`, `/v1/auth/login`,
    `/v1/auth/logout` and `/v1/shares/{token}` requires one of those. A share
    link is the credential for that one asset.

    **CORS.** Browsers may call the API only from origins listed in
    `CORS_ORIGIN` (comma-separated). The default is
    `http://localhost:5173,http://localhost:8080`.

    **Ownership.** A user sees assets they own and assets a team they belong
    to was granted. Someone else's asset answers `404`, exactly like a missing
    one, so ids cannot be probed. Admins use the `/v1/admin` routes to see
...
    once and stored only as a hash.

...

...

...
servers:
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /v1/assets/{assetId}/sharing:
    get:
      operationId: listAssetSharing
      tags: [Assets]
      summary: Teams and links for an asset the caller owns
      parameters:
        - $ref: '#/components/parameters/AssetId'
      responses:
        '200':
          description: Grants and links. The raw token is not included.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssetSharing' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/assets/{assetId}/shares:
    post:
      operationId: createShareLink
      tags: [Assets]
      summary: Create a share link
      description: |
        The raw token is returned once. Expiry must be in the future and within
        30 days. Download is off unless `canDownload` is true.
      parameters:
        - $ref: '#/components/parameters/AssetId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateShare' }
      responses:
        '201':
          description: The link, including the token this one time.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatedShare' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/shares/{token}:
    get:
      operationId: previewShare
      tags: [Shares]
      summary: Preview an asset with a share link
      security: []
      parameters:
        - $ref: '#/components/parameters/ShareToken'
      responses:
        '200':
          description: Filename, whether download is allowed, and a short-lived view.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SharePreview' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/shares/{token}/download:
    post:
      operationId: downloadShare
      tags: [Shares]
      summary: Download through a share link
      description: Counts as a download. Refused when the link does not allow it.
      security: []
      parameters:
        - $ref: '#/components/parameters/ShareToken'
      responses:
        '200':
          description: A short-lived download URL.
          content:
            application/json:
              schema:
                type: object
                required: [url]
                properties:
                  url: { type: string, format: uri }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: dam_token
      description: |
        HttpOnly session cookie set by register and login. Expires with the
        token (one hour by default). Not readable by page scripts.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Same token, for clients that are not a browser. The web app does not
        use this. Expires after one hour by default.
  parameters:
    AssetId:
      name: assetId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ShareToken:
      name: token
      in: path
      required: true
      description: The raw share token. It is not written to request logs.
      schema: { type: string, minLength: 20, maxLength: 128 }
...
  responses:
    ValidationError:
      description: |
        The request did not pass validation. A body or query problem lists each
        field. A value the database rejects for length, format or a check
        returns the same code with a generic message and no internal detail.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Unauthorized:
      description: Missing, invalid or expired token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: No such resource, or it belongs to someone else.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum:
                - validation_error
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - too_many_requests
                - service_unavailable
                - internal_error
            message: { type: string }
            requestId: { type: string }
...
    AssetStatus:
      type: string
      description: |
        `uploading` parts are still being sent; `uploaded` the file is stored
        and waiting for a worker; `processing` a worker has it; `ready` done;
        `failed` see `failureReason`.
      enum: [uploading, uploaded, processing, ready, failed]
...
    AssetSummary:
      type: object
      required:
        [
          id,
          filename,
          mimeType,
          sizeBytes,
          status,
          tags,
          createdAt,
          thumbnailUrl,
        ]
      properties:
        id: { type: string, format: uuid }
        filename: { type: string, maxLength: 255 }
        mimeType: { type: string, maxLength: 255 }
        sizeBytes: { type: integer, format: int64 }
        status: { $ref: '#/components/schemas/AssetStatus' }
        tags: { type: array, items: { type: string } }
        createdAt: { type: string, format: date-time }
        thumbnailUrl:
          type: [string, 'null']
          format: uri
          description: Presigned link, null until a thumbnail exists.
        access:
          type: string
          enum: [owner, team]
          description: |
            Present on the gallery list. `owner` can grant teams and create
            links. `team` can preview and download.
...
    AssetView:
      oneOf:
        - type: object
          required: [kind, renditions]
          properties:
            kind: { type: string, const: video }
            renditions:
              type: array
              items:
                type: object
                required: [label, width, height, url]
                properties:
                  label: { type: string, examples: ['1080p', '720p', original] }
                  width: { type: [integer, 'null'] }
                  height: { type: [integer, 'null'] }
                  url: { type: string, format: uri }
        - type: object
          required: [kind, url]
          properties:
            kind: { type: string, enum: [image, document] }
            url: { type: string, format: uri }
...
    AssetGrant:
      type: object
      required: [teamId, teamName, createdAt]
      properties:
        teamId: { type: string, format: uuid }
        teamName: { type: string }
        createdAt: { type: string, format: date-time }
    ShareLinkSummary:
      type: object
      required: [id, canDownload, expiresAt, revokedAt, createdAt]
      properties:
        id: { type: string, format: uuid }
        canDownload: { type: boolean }
        expiresAt: { type: string, format: date-time }
        revokedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
    AssetSharing:
      type: object
      required: [teams, links]
      properties:
        teams:
          type: array
          items: { $ref: '#/components/schemas/AssetGrant' }
        links:
          type: array
          items: { $ref: '#/components/schemas/ShareLinkSummary' }
    CreateShare:
      type: object
      required: [expiresAt]
      properties:
        canDownload: { type: boolean, default: false }
        expiresAt:
          type: string
          format: date-time
          description: Must be in the future and at most 30 days ahead.
    CreatedShare:
      type: object
      required: [id, token, canDownload, expiresAt]
      properties:
        id: { type: string, format: uuid }
        token:
          type: string
          description: Shown once. Later responses only keep a hash.
        canDownload: { type: boolean }
        expiresAt: { type: string, format: date-time }
    SharePreview:
      type: object
      required: [filename, mimeType, canDownload, expiresAt, asset, view]
      properties:
        filename: { type: string }
        mimeType: { type: string }
        canDownload: { type: boolean }
        expiresAt: { type: string, format: date-time }
        asset: { $ref: '#/components/schemas/AssetSummary' }
        view: { $ref: '#/components/schemas/AssetView' }